Add a version of comment with the profile picture to the left.
Add a costum route for CompletePost.

Show the comments with the most likes at the top.

Add a way to navigate to the post by clicking it.

Add a way to navigate to the user by clicking the profile picture in post and comments.

Add hover effects to the comment and like buttons.

Add a component (TendencyUsers.cs) that shows the users with most likes with a costum amount.