TendencyUsers.cshtml -Tried to make my own sorting algorithm to beat OrderByDescending but i was slower.
-Move the counting of likes to a new function and added a Span loop.

-Create the method GetUsersFromKeyValuePair
 to get the ordered users from the KeyValuePair with a span loop.

-Add a span loop everywhere i could.

Decrease the padding in like and comment button and also add a title.

Change DeletePost removes the post, now its handle by RemovePostFromContainer.

DeletePost now accepts the username a parameter, now we don't to get the username from the url.

Add a stopPropagation to comments, now clicking the comments doesn't active redirect.

Fix UpdatePostFromContainer problem where it nedded the username of theuser to get the post container by adding a data attribute to the posts that defines the username. It can also be use in CompletePost now.

Deleting a post in CompletePost now redirects you back to you UserPage. Its done a function only defined in CompletePost.