Select only the necessary properties for the post an users in TendencyUsers (massive improvement).
Improve performance and memory allocation of GetPosts and Load Comments by:

-GetPostLikesSelective loads only the ids of the users who like a post and returns a list of users with only the id property. The comment variation is GetCommentLikesSelective.

-LoadComments now loads only the necessary propeties of the user of a comment.

Now the article element that contains the post and his comments has the id of the post. There is no need to pass the userName to find the post container.

When a comment is created/deleted the amount of comments in post-stats is updated.

Move the remove event to Background in PostScript.

Create a tab that shows the users who a liked a post or comment.