Refactor some PostController action methods: -CreatePost, initiate UserLikes and Comments properties with empty list and use only one model to pass it to the partial view.
-EditPost, use only on model to pass it to the partial view and update the methods to the likes and comments.

-CreateComment, refactor the code to follow the same pattern as CreatePost.

-LikePost, refactor the code to follow the same pattern as LikePost.

Add a page that shows the liked post of an user. LikedPosts is the action method and the name of the view.

Add a partial view "PageUserNav" that acts as the navigation bar to the user' posts, comments,
liked posts, etc.

Add a partial view "PostsList" that acepts a list of Post Model and shows the post and its comments.

Add a page that shows the posts commented by an user. CommentedPosts is the action method and the name of the view.