In EditPostTab change get method for an ajax one. This was done so that i could display error messages to the user sended by LookForPost when the post does not exist or does not belong to the user.
Add an autofocus to the Create Post / Comment and EditPost tabs.

Adjust UserPage to show a message when user does not exist.

Adjust CompletePost to show a post not found page when post does not exist or does not belong to the user of the page.

Fix error in CompletePost where the post was show even when it did not belong to the user of the page.

Now Posts are loaded as the client scrolls down.

GetPostsNew and GetLikedPostNew, fetch posts from the database using a range (from, to).

KeepLoadingPosts  (UserController.cs), returns a list of posts in an html text using the partial view PostList.

KeepLoadingPosts and LoadMorePostsMedia  (PostScript.js), makes the call passing the id of the user, from where to start and where to end. Also passes the response to AddRangePosts.

AddRangePosts, converts the posts text in html objects and appends them to the main container.

SetScrollEvent, add the scroll event to the window as well as containing the paramaters for future calls. The event only gets trigger when the users scrolls an exact percentaje of the main container (center-box).