Delete AllUsers.cs component.
Fix duplicated post in GetCommentedPosts by loading all the comments of a post and canceling a post if it was already loaded.

Add a singleton service that contains users and posts stats (most likes, posts, comments and more in the future).

Add partial view Right-Box that contains the HTML of the right-box class in various view.

Delete TendencyUsers.cs component.

Add blur to the media when showImages propertie is true