Add an explore page that load random posts.
Explore.cs service loads the random posts, the posts are not duplicated, although the list of posts is resete when redirecting.

Change the onload event in Tendency.js for an addEventListener.

Add the small-user-list to make UsersList.cs fit in left and right box.

Make right-box sticky with hidden scroll funtionability.

Change how a complete post is show, now it loads a partial view intead of redirecting to CompletePost