Add a tabs to explore page that loads the most recent post and the oldest post.
Add a visual indicator en explore page to let the user know in which tab is on.

Refactor the code in Explore.js, now swithing between is handle by a single method (SwitchToTab) and loading more post when srolling is also handle by a single method (LoadMorePostWhenScrolling).

Refactor some of the code in ExploreController.cs by adding FillPostsProperties and LoadPartialViewInfo.

Refactor and optimize the code in UserController.cs when loading post's properties (UserLikes, User, Comments) using the method FillPostsProperties. LoadPostListInfo now takes care of the partial view info.

Some changes are not done.