Change the list of strings for a StringBuilder in FillPostsProperties and NewLoadComments.
Move evething related to fetching data from the database and putting it in posts and comments properties to a transient service ModelLogic. The methods are FillPostsProperties, RangeLoadComments, GetPostLikesSelective and GetCommentLikesSelective.

Update AddPostToContainer to the new way the id is store and add the ShowCompletePosts to the onclick event.

Create new methods in ModelLogic when only 1 model is provided to load one or more properties. Methods : SingleFillPostProperties and SingleLoadComments.

Fix the width of a post when show in the view post tab.

FIx the autofocus and put the cursor at the end in edit post tab.

Change the padding an resize the close button in the view post tab to look better on mobile.

Modify the style of tab-list and the settings icon to be more mobile friendly.