Delete the HasDefaultValue value for EditedDate.
Remove Comments, LikedPost and LikedComments properties from UserModel.

Delete database and migrations.

Add an appropiate datatype to some properties of models in fluenAPI.

Renamed PostDate to Date in PostModel.

Change userModel table name from AspNetUsers to Users.

Delete hipsters content from posts and coments.

Create a random age if age its too big for NormalUser.

Add new models PostLikes and CommentLikes, to save the likes in the database. Similar to the old likes schema but now with the date the user like a post or comment (LikedDate).

New like schema working.

Fix the order in LikedPosts, now is ordered by when the user like the post.

Add new model FollowUsers that handles the relationship of followers.

Add a follow button to follow or unfollown a user.

Add methods in ModelLogic, GetFollowers for returning the followers ids of an user and GetFollowingUsers for the users an user following.

Show the amount of followers and following users in PageUserInfo.