change glodbal css variables names.
Following an user now updates the amount of followers in the left-box as well as in the userDisplayTab.

Rename Tendency js file to UserScript.

Change of the methods that load more posts for PostsScript and the UserController.

Now the amount of posts loaded when scrolling down is centralized by a constant in UserController.

Refactor SwitchToPosts and SwichToComments into new method SwitchTabLikedPostsAndComments.

Delete UserDisplayTab partial view, just put its code in WebLayout.

Create tabs to show the followers and following users.

Add a check in Follow method so a user can not follow himself.

Add data attibutes to user stats info to use in UpdateUserStats.