UpdateUserStats method created in UserController, returns page user stats.
In PageUserInfo add the id of the user in a "hidden"element.

UpdateUserStats new method in UserScript, calls UpdateUserStats in the UserController. After the result is return updates the DOM of PageUserInfo (for the left-box and the display-tab). Gets call every 10 second, after creating a post, after liking a post, after removing a post, after updating a post, creating a comment, liking a comment, deleting a comment and following/ unfollowing a user.

Delete AlterFollowers method in UserScript, replaced by UpdateUserStats.

Create new background service (RandomFollowers) to add followers to random users.

Change the querry in FollowersUsersTab and FollowingUsersTab to order users by the date they followed.

Add a list of Users with most followers in Tendency.cs and show the list in the right-box.

Create new page call "Home" that loads all the following user's posts of the logged user.