Refactor all of the LoadMore... methods in PostScipt into one, LoadMorePostsGlobal now handles the fetching of more posts when a user scrolls down.
New "enum", actionMethods holds the names of all the LoadMore... of the UserController. Now changing the names is way easier.

Move the SetScrollEvent and loadingPosts to PostScript.

UpdatePostInfo new method in PostController.cs returns the amount of likes and comments of a specific post in json format.

UpdatePostInfo new method in PostScript.js makes a call to UpdatePostInfo in PostController.cs  and uses the reponse to update the post in the DOM.

New global js variable postObserver of type IntersectionObserver.

New feature, now posts will update automatically the amount of likes and comments every x seconds only if they are currently on the viewport.

Make left box and display tab share the same user info element. The element changes parent using a media query on MobileTabs.

Make the setting anchor a tab list that displays all the pages in the settings controller and finally add a log out button.

Change the name of the model FollowUsers to Followers and its table name too.

Detele the database, this seems to have fex the bug with images in posts.