Add messages in case the Posts properties does not have a single post.
Delete ImagePrev.js for page that don't use it.

Add a method InitialSeed to all the sevices to fill the database when is empty.

Fix bug in RandomPosts where it will stop because of an error.

Add a sparkling effect when liking a post and when the UpdatePostInfo increases the amount of likes and comments of a post..

Move the hover effects of post-stats buttons to desktop only (screen width > 768px).

Add a custom scroll bar to the post's content.

Add new action methods to explore controller for the top, recent and old tabs. Also change the url when changing tabs.

Add enums to idenitfy teh tabs.