Move the selection of the scroll event to Explore.js.
Create an "enum" for the names of the action methods and the tab names.

Change ViewDatas to ViewBags in explorer controller an its view.

Pass the amountPostsPerLoad to the view so for future changes just need to change that one variable.

Add parameters checks to all the methods in Explore.js.

Add or improve the parameters checks to the functions in PostScript.js.

Add a function in PostScript to check for strings, IsString that returns true if a string is pass.

Finnaly delete DynamicUser model, there was not a need for it in the first place.

Delete UserPage, MediaPosts, LikedPostsAndComments, CommentedPosts and replace them all for Profile view. Also made the respective adjustments to the PostScript.js file.

Refactor LoadPageUserStats.

Delete CompletePost action method and view, no longer needed.

Delete SearchUser action method, not needed at all.

Delete PageUserNav and put its html in Profile.