Update Home.cshtml with the new SetScrollEvent.
Create a new class center-sparkling-like, for a version of sparkling more centered. To use it just pass true as the second argument in SparklingLikeEffect.

Add the sparkling effect when updating a user stats it goes up.

Create a new controller, the SearchController  and move SearchUsers action method and view there. Also move other action methods as UserController is quite stuff.

Create action methods in the searh controllet for getting the users with most likes, followers, etc.

Could not finnish had to stop for the day.